+++ internal/web/static/app.js
@@ -18,6 +18,7 @@ import { focusPane } from "./keys.js";
 import { fetchHealth } from "./bigrepo.js";
 import "./settings.js";
 import "./identity.js";
+import "./prefixes.js";
 import "./palette.js";
 
 async function boot() {